# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Gestion de Comptes Bancaires",
        "description": "Documentation complète de l'API REST pour la gestion des comptes bancaires. Cette API permet de gérer les comptes clients avec pagination, filtres avancés et recherche.",
        "contact": {
            "name": "Support API",
            "email": "support@example.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8001/api/v1",
            "description": "Serveur de développement local"
        },
        {
            "url": "http://localhost:8001/api/v1",
            "description": "Serveur local alternatif"
        }
    ],
    "paths": {
        "/comptes": {
            "get": {
                "tags": [
                    "Comptes"
                ],
                "summary": "Lister tous les comptes",
                "description": "Récupère la liste de tous les comptes avec pagination et filtres",
                "operationId": "getComptes",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Numéro de page",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 1
                        }
                    },
                    {
                        "name": "limit",
                        "in": "query",
                        "description": "Nombre d'éléments par page",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 10,
                            "maximum": 100
                        }
                    },
                    {
                        "name": "type",
                        "in": "query",
                        "description": "Filtrer par type",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "epargne",
                                "cheque"
                            ]
                        }
                    },
                    {
                        "name": "statut",
                        "in": "query",
                        "description": "Filtrer par statut",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "actif",
                                "bloque",
                                "ferme"
                            ]
                        }
                    },
                    {
                        "name": "search",
                        "in": "query",
                        "description": "Recherche par titulaire ou numéro",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "sort",
                        "in": "query",
                        "description": "Tri",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "default": "dateCreation",
                            "enum": [
                                "dateCreation",
                                "derniereModification",
                                "numeroCompte"
                            ]
                        }
                    },
                    {
                        "name": "order",
                        "in": "query",
                        "description": "Ordre",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "default": "desc",
                            "enum": [
                                "asc",
                                "desc"
                            ]
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Liste des comptes récupérée avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Liste des comptes récupérée avec succès"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Compte"
                                            }
                                        },
                                        "pagination": {
                                            "type": "object"
                                        },
                                        "links": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Non autorisé"
                    },
                    "422": {
                        "description": "Erreur de validation"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/comptes/{id}": {
            "get": {
                "tags": [
                    "Comptes"
                ],
                "summary": "Obtenir un compte par ID",
                "description": "Récupère les détails d'un compte spécifique",
                "operationId": "getCompteById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID du compte",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Compte récupéré avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Compte récupéré avec succès"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/Compte"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Compte non trouvé"
                    },
                    "401": {
                        "description": "Non autorisé"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/comptes/numero/{numero}": {
            "get": {
                "tags": [
                    "Comptes"
                ],
                "summary": "Obtenir un compte par numéro",
                "description": "Récupère les détails d'un compte par son numéro",
                "operationId": "getCompteByNumero",
                "parameters": [
                    {
                        "name": "numero",
                        "in": "path",
                        "description": "Numéro du compte",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Compte récupéré avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Compte récupéré avec succès"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/Compte"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Compte non trouvé"
                    },
                    "401": {
                        "description": "Non autorisé"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/health": {
            "get": {
                "tags": [
                    "Health"
                ],
                "summary": "Vérifier l'état de l'API",
                "description": "Endpoint pour vérifier si l'API fonctionne correctement",
                "operationId": "healthCheck",
                "responses": {
                    "200": {
                        "description": "L'API fonctionne correctement",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "API is running"
                                        },
                                        "version": {
                                            "type": "string",
                                            "example": "v1"
                                        },
                                        "timestamp": {
                                            "type": "string",
                                            "format": "date-time",
                                            "example": "2025-10-23T12:29:16+00:00"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "ErrorResponse": {
                "title": "Réponse d'erreur",
                "description": "Format standard pour les réponses d'erreur",
                "properties": {
                    "success": {
                        "type": "boolean",
                        "example": false
                    },
                    "message": {
                        "type": "string",
                        "example": "Erreur lors du traitement de la requête"
                    },
                    "errors": {
                        "description": "Détails des erreurs de validation (optionnel)",
                        "type": "object"
                    }
                },
                "type": "object"
            },
            "PaginationMeta": {
                "title": "Métadonnées de pagination",
                "description": "Informations sur la pagination des résultats",
                "properties": {
                    "currentPage": {
                        "description": "Page actuelle",
                        "type": "integer",
                        "example": 1
                    },
                    "totalPages": {
                        "description": "Nombre total de pages",
                        "type": "integer",
                        "example": 5
                    },
                    "totalItems": {
                        "description": "Nombre total d'éléments",
                        "type": "integer",
                        "example": 50
                    },
                    "itemsPerPage": {
                        "description": "Nombre d'éléments par page",
                        "type": "integer",
                        "example": 10
                    },
                    "hasNext": {
                        "description": "Indique s'il y a une page suivante",
                        "type": "boolean",
                        "example": true
                    },
                    "hasPrevious": {
                        "description": "Indique s'il y a une page précédente",
                        "type": "boolean",
                        "example": false
                    },
                    "from": {
                        "description": "Index du premier élément de la page",
                        "type": "integer",
                        "example": 1
                    },
                    "to": {
                        "description": "Index du dernier élément de la page",
                        "type": "integer",
                        "example": 10
                    }
                },
                "type": "object"
            },
            "PaginationLinks": {
                "title": "Liens de pagination",
                "description": "Liens de navigation pour la pagination",
                "properties": {
                    "self": {
                        "description": "Lien vers la page actuelle",
                        "type": "string",
                        "example": "/api/v1/comptes?page=1"
                    },
                    "first": {
                        "description": "Lien vers la première page",
                        "type": "string",
                        "example": "/api/v1/comptes?page=1"
                    },
                    "last": {
                        "description": "Lien vers la dernière page",
                        "type": "string",
                        "example": "/api/v1/comptes?page=5"
                    },
                    "next": {
                        "description": "Lien vers la page suivante",
                        "type": "string",
                        "example": "/api/v1/comptes?page=2",
                        "nullable": true
                    },
                    "previous": {
                        "description": "Lien vers la page précédente",
                        "type": "string",
                        "example": null,
                        "nullable": true
                    }
                },
                "type": "object"
            },
            "Compte": {
                "title": "Compte",
                "description": "Modèle de compte bancaire",
                "properties": {
                    "id": {
                        "description": "Identifiant unique du compte",
                        "type": "string",
                        "format": "uuid",
                        "example": "a02ea57f-907e-4894-acab-de01af9d4163"
                    },
                    "numeroCompte": {
                        "description": "Numéro unique du compte (format: CP + 10 chiffres)",
                        "type": "string",
                        "maxLength": 20,
                        "example": "CP0241262525"
                    },
                    "titulaire": {
                        "description": "Nom complet du titulaire du compte",
                        "type": "string",
                        "example": "Prof. Reta Lesch"
                    },
                    "type": {
                        "description": "Type de compte",
                        "type": "string",
                        "enum": [
                            "cheque",
                            "epargne"
                        ],
                        "example": "cheque"
                    },
                    "devise": {
                        "description": "Devise du compte",
                        "type": "string",
                        "default": "FCFA",
                        "example": "FCFA"
                    },
                    "dateCreation": {
                        "description": "Date de création du compte",
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-10-23T12:20:55+00:00"
                    },
                    "statut": {
                        "description": "Statut du compte",
                        "type": "string",
                        "enum": [
                            "actif",
                            "bloque",
                            "ferme"
                        ],
                        "example": "actif"
                    },
                    "motifBlocage": {
                        "description": "Motif du blocage si le compte est bloqué",
                        "type": "string",
                        "example": null,
                        "nullable": true
                    },
                    "metadata": {
                        "description": "Métadonnées du compte",
                        "properties": {
                            "derniereModification": {
                                "description": "Date de dernière modification",
                                "type": "string",
                                "format": "date-time",
                                "example": "2025-10-23T12:20:55+00:00"
                            },
                            "version": {
                                "description": "Version pour le contrôle de concurrence optimiste",
                                "type": "integer",
                                "example": 1
                            }
                        },
                        "type": "object"
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Authentification via Laravel Sanctum. Pour obtenir un token, utilisez l'endpoint /auth/login (actuellement désactivé en développement)",
                "bearerFormat": "Token",
                "scheme": "bearer"
            }
        }
    },
    "tags": [
        {
            "name": "Comptes",
            "description": "Endpoints pour la gestion des comptes bancaires - Création, consultation, modification et suppression"
        },
        {
            "name": "Health",
            "description": "Endpoints de vérification de l'état de l'API"
        }
    ]
}